[partial] Fix bug LeafletLayer
LeafletPath (LeafletLayer) is behaving unconsistently sometimes, not destroying the map layer or not displaying it.

Using GeoJSON name reather than a (pseudo)random string as layer identifier seems to work better.

Display issue is (probably) due to flyToBounds when bounds are actualy the same as set. Paning map seems to solve the issue [sic].

[!] Destroy issue need more investigation and tests